chore(core-client): reshare and backup to own files
